requirements.txt
revolt.py==0.2.0
aiohttp
